refactor(scripts): extract templates from create-new-lib/app generators
The two scaffolding generators were 1618 and 1317 lines respectively because
every template file was inlined as a JS string. Editing a single field meant
hunting through thousands of escaped backticks; templates weren't lint-checked
or type-checked; the two scripts had drifted (workspace registration,
prompts, file writing all duplicated).

Move templates into real files under scripts/templates/{lib,app}/ with
{{PLACEHOLDER}} substitution, and share workspace + Dockerfile registration
via scripts/lib/template-engine.mjs. The orchestrator scripts now only
parse args, render templates, and patch package.json for variant-specific
deps.

- create-new-lib.js: 1618 -> 212 lines (87% reduction)
- create-new-app.js: 1317 -> 196 lines (85% reduction)
- shared engine:     0    -> 132 lines

Output parity verified byte-for-byte against the legacy generator for all
6 variants: lib (service / utility / service+api) and app (minimal /
standard / enterprise). Generated libs pass check-lib-tests.mjs and the
service variant's vitest suite continues to pass; the with-api test
failure is a pre-existing issue in the original template (unchanged).

Templates contain {{PLACEHOLDER}} tokens that aren't valid TS/JSON, so
scripts/templates/ is added to both .prettierignore and the root eslint
config's ignores list.
