+++ scripts/templates/app/enterprise/package.json
@@ -1,0 +1,55 @@
+{
+  "name": "@adopt-dont-shop/{{APP_NAME}}",
+  "version": "1.0.0",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --coverage",
+    "lint": "eslint src --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "lint:fix": "eslint src --ext ts,tsx --fix",
+    "type-check": "tsc --noEmit"
+  },
+  "dependencies": {
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.8.1",
+    "@adopt-dont-shop/lib.components": "*",
+    "@adopt-dont-shop/lib.auth": "*",
+    "@adopt-dont-shop/lib.analytics": "*",
+    "@adopt-dont-shop/lib.api": "*",
+    "@adopt-dont-shop/lib.feature-flags": "*",
+    "@adopt-dont-shop/lib.notifications": "*",
+    "@adopt-dont-shop/lib.permissions": "*",
+    "@adopt-dont-shop/lib.discovery": "*",
+    "@adopt-dont-shop/lib.search": "*",
+    "@statsig/react-bindings": "^3.18.2",
+    "@tanstack/react-query": "^5.0.0",
+    "styled-components": "^6.1.8"
+  },
+  "devDependencies": {
+    "@types/jest": "^29.4.0",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
+    "@vitejs/plugin-react": "^4.3.4",
+    "eslint": "^8.45.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.3",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^29.4.0",
+    "jest-environment-jsdom": "^29.4.0",
+    "jest-transform-stub": "^2.0.0",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.0.2",
+    "vite": "^5.4.10",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/jest-dom": "^6.0.0",
+    "@testing-library/user-event": "^14.0.0"
+  }
+}
