+++ scripts/templates/app/enterprise/src/contexts/NotificationsContext.tsx
@@ -1,0 +1,61 @@
+import { NotificationsService, Notification } from '@adopt-dont-shop/lib.notifications';
+import { createContext, useContext, ReactNode, useMemo, useState } from 'react';
+
+interface NotificationsContextType {
+  notificationsService: NotificationsService;
+  notifications: Notification[];
+  unreadCount: number;
+  markAsRead: (id: string) => Promise<void>;
+  isLoading: boolean;
+}
+
+const NotificationsContext = createContext<NotificationsContextType | null>(null);
+
+export const useNotifications = () => {
+  const context = useContext(NotificationsContext);
+  if (!context) {
+    throw new Error('useNotifications must be used within NotificationsProvider');
+  }
+  return context;
+};
+
+interface NotificationsProviderProps {
+  children: ReactNode;
+}
+
+export const NotificationsProvider = ({ children }: NotificationsProviderProps) => {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+  const [isLoading] = useState(false);
+  
+  const notificationsService = useMemo(() => {
+    return new NotificationsService({
+      apiUrl: import.meta.env.VITE_API_BASE_URL,
+      debug: import.meta.env.NODE_ENV === 'development'
+    });
+  }, []);
+
+  const unreadCount = useMemo(() => {
+    return notifications.filter(n => !n.readAt).length;
+  }, [notifications]);
+
+  const markAsRead = async (id: string) => {
+    // Simplified implementation
+    setNotifications(prev => 
+      prev.map(n => n.id === id ? { ...n, readAt: new Date().toISOString() } : n)
+    );
+  };
+
+  const value = useMemo(() => ({
+    notificationsService,
+    notifications,
+    unreadCount,
+    markAsRead,
+    isLoading,
+  }), [notificationsService, notifications, unreadCount, isLoading]);
+
+  return (
+    <NotificationsContext.Provider value={value}>
+      {children}
+    </NotificationsContext.Provider>
+  );
+};
