+++ scripts/templates/app/minimal/src/App.tsx
@@ -1,0 +1,20 @@
+import { Route, Routes } from 'react-router-dom';
+import { HomePage } from '@/pages/HomePage';
+import { DevLoginPanel } from '@/components/dev/DevLoginPanel';
+import { Footer } from '@adopt-dont-shop/lib.components';
+
+function App() {
+  return (
+    <div className="app">
+  <main>
+    <Routes>
+      <Route path="/" element={<HomePage />} />
+    </Routes>
+  </main>
+  <DevLoginPanel />
+  <Footer />
+</div>
+  );
+}
+
+export default App;
