+++ scripts/templates/app/minimal/src/main.tsx
@@ -1,0 +1,18 @@
+import { ThemeProvider } from '@adopt-dont-shop/lib.components';
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import { BrowserRouter } from 'react-router-dom';
+import App from './App';
+import { AuthProvider } from '@/contexts/AuthContext';
+
+ReactDOM.createRoot(document.getElementById('root')!).render(
+  <React.StrictMode>
+    <AuthProvider>
+      <ThemeProvider>
+        <BrowserRouter>
+          <App />
+        </BrowserRouter>
+      </ThemeProvider>
+    </AuthProvider>
+  </React.StrictMode>
+);
