+++ scripts/templates/lib/common/eslint.config.js
@@ -1,0 +1,3 @@
+import baseConfig from '@adopt-dont-shop/eslint-config-base';
+
+export default baseConfig;
