+++ scripts/templates/lib/common/vitest.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig, mergeConfig } from 'vitest/config';
+import sharedConfig from '../vitest.shared.config';
+
+export default mergeConfig(
+  sharedConfig,
+  defineConfig({
+    test: {
+      name: 'lib.{{LIB_NAME}}',
+      setupFiles: ['./src/test-utils/setup-tests.ts'],
+    },
+  })
+);
