+++ scripts/templates/lib/service/src/index.ts
@@ -1,0 +1,7 @@
+// Main exports for @adopt-dont-shop/lib-{{LIB_NAME}}
+export { {{SERVICE_NAME}} } from './services/{{LIB_NAME}}-service';
+export type {
+  {{SERVICE_NAME}}Config,
+  {{SERVICE_NAME}}Options,
+} from './types';
+export * from './types';
