+++ scripts/templates/lib/service/src/services/service.ts
@@ -1,0 +1,85 @@
+import { {{SERVICE_NAME}}Config, {{SERVICE_NAME}}Options } from '../types';
+
+/**
+ * {{SERVICE_NAME}} - Handles {{LIB_NAME}} operations
+ */
+export class {{SERVICE_NAME}} {
+  private config: {{SERVICE_NAME}}Config;
+
+  constructor(config: Partial<{{SERVICE_NAME}}Config> = {}) {
+    this.config = {
+      debug: false,
+      ...config,
+    };
+
+    if (this.config.debug) {
+      console.log(`${{{SERVICE_NAME}}.name} initialized with config:`, this.config);
+    }
+  }
+
+  /**
+   * Get current configuration
+   */
+  public getConfig(): {{SERVICE_NAME}}Config {
+    return { ...this.config };
+  }
+
+  /**
+   * Update service configuration
+   */
+  public updateConfig(updates: Partial<{{SERVICE_NAME}}Config>): void {
+    this.config = { ...this.config, ...updates };
+    
+    if (this.config.debug) {
+      console.log(`${{{SERVICE_NAME}}.name} config updated:`, this.config);
+    }
+  }
+
+  /**
+   * Example method - customize based on your library's purpose
+   * TODO: Replace with actual {{LIB_NAME}} functionality
+   */
+  public async exampleMethod(
+    data: Record<string, unknown>,
+    _options: {{SERVICE_NAME}}Options = {}
+  ): Promise<{ success: boolean; data: any; message?: string }> {
+    try {
+      if (this.config.debug) {
+        console.log(`${{{SERVICE_NAME}}.name} exampleMethod called with:`, data);
+      }
+
+      // TODO: Implement your actual logic here
+      const result = {
+        success: true,
+        data: { processed: data, timestamp: new Date().toISOString() },
+        message: 'Example operation completed successfully',
+      };
+
+      if (this.config.debug) {
+        console.log(`${{{SERVICE_NAME}}.name} exampleMethod completed`);
+      }
+
+      return result;
+    } catch (error) {
+      if (this.config.debug) {
+        console.error(`${{{SERVICE_NAME}}.name} exampleMethod failed:`, error);
+      }
+      throw error;
+    }
+  }
+
+  /**
+   * Health check method
+   */
+  public async healthCheck(): Promise<boolean> {
+    try {
+      // TODO: Implement actual health check logic
+      return true;
+    } catch (error) {
+      if (this.config.debug) {
+        console.error(`${{{SERVICE_NAME}}.name} health check failed:`, error);
+      }
+      return false;
+    }
+  }
+}
