+++ scripts/templates/lib/utility/package.json
@@ -1,0 +1,73 @@
+{
+  "name": "@adopt-dont-shop/lib-{{LIB_NAME}}",
+  "version": "1.0.0",
+  "description": "{{LIB_DESCRIPTION}}",
+  "type": "module",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist",
+    "README.md"
+  ],
+  "scripts": {
+    "build": "tsc",
+    "dev": "tsc --watch",
+    "clean": "rm -rf dist",
+    "test": "vitest run",
+    "test:watch": "vitest",
+    "test:coverage": "vitest run --coverage",
+    "lint": "eslint src",
+    "lint:fix": "eslint src --fix",
+    "type-check": "tsc --noEmit",
+    "prepublishOnly": "npm run clean && npm run build"
+  },
+  "keywords": [
+    "pet-adoption",
+    "library",
+    "typescript",
+    "react",
+    "components"
+  ],
+  "author": "Adopt Don't Shop Team",
+  "license": "MIT",
+  "dependencies": {
+    "@types/node": "^20.19.0",
+    "react": "^18.2.0",
+    "styled-components": "^6.1.8"
+  },
+  "devDependencies": {
+    "@adopt-dont-shop/eslint-config-base": "*",
+    "@typescript-eslint/eslint-plugin": "^8.59.1",
+    "@typescript-eslint/parser": "^8.59.1",
+    "@types/react": "^18.2.0",
+    "@types/styled-components": "^5.1.26",
+    "eslint": "^9.0.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "prettier": "^3.2.5",
+    "typescript": "^5.4.5",
+    "typescript-eslint": "^8.0.0",
+    "vitest": "^4.0.8",
+    "@vitest/coverage-v8": "^4.0.8"
+  },
+  "peerDependencies": {
+    "typescript": "^5.0.0",
+    "react": ">=18.0.0",
+    "styled-components": ">=5.0.0"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ParagonJenko/pet-adoption-react.git",
+    "directory": "lib.{{LIB_NAME}}"
+  },
+  "bugs": {
+    "url": "https://github.com/ParagonJenko/pet-adoption-react/issues"
+  },
+  "homepage": "https://github.com/ParagonJenko/pet-adoption-react/tree/main/lib.{{LIB_NAME}}#readme"
+}
