+++ scripts/templates/lib/utility/src/components/index.ts
@@ -1,0 +1,10 @@
+// Components for @adopt-dont-shop/lib-{{LIB_NAME}}
+
+// Export individual components here as you create them
+// Example:
+// export { MyComponent } from './MyComponent';
+
+// Placeholder component - replace with actual components
+export const {{PASCAL_CASE_NAME}}Component = () => {
+  return null; // Replace with actual component implementation
+};
