+++ scripts/templates/lib/utility/src/hooks/index.ts
@@ -1,0 +1,11 @@
+// Hooks for @adopt-dont-shop/lib-{{LIB_NAME}}
+
+// Export individual hooks here as you create them
+// Example:
+// export { useMyHook } from './useMyHook';
+
+// Placeholder hook - replace with actual hooks
+export const use{{PASCAL_CASE_NAME}} = () => {
+  // Replace with actual hook implementation
+  return {};
+};
