+++ scripts/templates/lib/utility/src/index.ts
@@ -1,0 +1,10 @@
+// Main exports for @adopt-dont-shop/lib-{{LIB_NAME}}
+
+// Re-export all components
+export * from './components';
+
+// Re-export all hooks  
+export * from './hooks';
+
+// Re-export all utilities
+export * from './utils';
