+++ scripts/templates/lib/utility/src/utils/index.ts
@@ -1,0 +1,12 @@
+// Utilities for @adopt-dont-shop/lib-{{LIB_NAME}}
+
+// Export individual utility functions here as you create them
+// Example:
+// export { myUtilFunction } from './myUtilFunction';
+
+/**
+ * Placeholder utility function - replace with actual utilities
+ */
+export const {{CAMEL_CASE_NAME}}Utils = {
+  // Add utility functions here
+};
